fix: avoid deleting non-existent files
Previously, the code attempted to delete a file without checking its existence, causing errors.
 This update adds a condition to ensure the file exists before deletion.

+++ src/org/omegat/core/segmentation/SRX.java
@@ -133,7 +133,9 @@ public class SRX implements Serializable {
         File outFile = new File(outDir, SRX_SENTSEG);
 
         if (srx == null) {
-            Files.delete(outFile.toPath());
+            if (outFile.exists()) {
+                Files.delete(outFile.toPath());
+            }
             return;
         }
